# Faster equality, faster hashing, only keep unique boards in beam search