**before**
```
Benchmarking: anytime_beam_search(board, {'power': 10})
 ..........
 Solutions of avg. length 16.200 in 27.834 sec.
 Avg. length times sec (lower is better) => 450.918

Benchmarking: heuristic_search(board, {'max_nodes': 220000})
 ..........
 Solutions of avg. length 17.600 in 34.846 sec.
 Avg. length times sec (lower is better) => 613.284

Benchmarking: monte_carlo_search(board, {'iterations': 3100, 'seed': 1})
 ..........
 Solutions of avg. length 16.500 in 33.637 sec.
 Avg. length times sec (lower is better) => 555.018
```


**after**
```
Benchmarking: anytime_beam_search(board, {'power': 10})
 ..........
 Solutions of avg. length 15.500 in 17.190 sec.
 Avg. length times sec (lower is better) => 266.448

Benchmarking: heuristic_search(board, {'max_nodes': 220000})
 ..........
 Solutions of avg. length 17.600 in 17.300 sec.
 Avg. length times sec (lower is better) => 304.472

Benchmarking: monte_carlo_search(board, {'iterations': 3100, 'seed': 1})
 ..........
 Solutions of avg. length 16.500 in 18.075 sec.
 Avg. length times sec (lower is better) => 298.234
```